+++ research/mlflow_experiments/utils/metrics.py
@@ -18,18 +18,18 @@ def compute_retrieval_metrics(ground_truth, predictions, k_values=None):
             hits = sum(1 for doc in top if doc in rel)
             p_scores.append(hits / k)
             r_scores.append(hits / len(rel))
-        metrics[f"mean_P@{k}"] = float(np.mean(p_scores)) if p_scores else 0.0
-        metrics[f"mean_R@{k}"] = float(np.mean(r_scores)) if r_scores else 0.0
+        metrics[f"mean_P@{k}"] = np.mean(p_scores) if p_scores else 0
+        metrics[f"mean_R@{k}"] = np.mean(r_scores) if r_scores else 0
 
     rr = []
     for qid, relevant in ground_truth.items():
         retrieved = predictions.get(qid)
         if not retrieved:
-            rr.append(0.0)
+            rr.append(0)
             continue
         rel = set(relevant)
         rank = next((i for i, doc in enumerate(retrieved, 1) if doc in rel), None)
-        rr.append(1.0 / rank if rank else 0.0)
-    metrics["MRR"] = float(np.mean(rr)) if rr else 0.0
+        rr.append(1 / rank if rank else 0)
+    metrics["MRR"] = np.mean(rr) if rr else 0
 
     return metrics
